test case for --as-required
